requirements.txt
# LLM & Embeddings
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.10
ollama>=0.1.6

# Vector Database
chromadb>=0.4.22

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
markdown>=3.5.0

# Data Processing
pandas>=2.1.0
numpy>=1.26.0,<2.0.0

# Experiment Tracking
mlflow>=2.9.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Multi-Agent Framework
langgraph>=0.0.20
langchain-experimental>=0.0.49

# SQL Support
sqlalchemy>=2.0.0

# Code Analysis
gitpython>=3.1.40

# Core dependency - compatible with guardrails-ai
pydantic>=2.0.0,<2.5.0

# Network and utilities
networkx>=3.2.0

# Evaluation Frameworks
ragas>=0.1.0
# deepeval>=0.20.0  # Optional - not used in current implementation

# Guardrails (requires pydantic<2.5)
guardrails-ai>=0.4.0,<0.5.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Observability
arize-phoenix>=3.0.0
openinference-instrumentation-langchain>=0.1.0

# Evaluation metrics
rouge-score>=0.1.2
spacy>=3.7.0

# Additional dependencies that might be needed
requests>=2.31.0
